Update tests and Carthage
