+++ Tests/FirebaseTrackerTests.swift
@@ -22,25 +22,11 @@ class FirebaseTrackerTests: XCTestCase {
 
     override func tearDown() { }
 
-    func createRemoteCommandResponse(type: SimpleCommandType, commandId: String, payload: [String: Any]) -> RemoteCommandResponseProtocol? {
-        switch type {
-        case .webview:
-            let responseDescription = HttpTestHelpers.httpRequestDescription(commandId: commandId, config: [:], payload: payload)
-            if let description = responseDescription {
-                return RemoteCommandResponse(urlString: description)
-            }
-        case .JSON:
-            return JSONRemoteCommandResponse(with: payload)
-        }
-        XCTFail("Could not create Remote Command Response description from stubs provided")
-        return nil
-    }
-
     // MARK: Webview Remote Command Tests
     func testCreateAnalyticsConfigWithoutValues() {
         let expect = expectation(description: "firebase config should run")
         let payload: [String: Any] = ["command_name": "config"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.createAnalyticsConfigCallCount)
         }
@@ -51,7 +37,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testCreateAnalyticsConfigWithValues() {
         let expect = expectation(description: "firebase config should run")
         let payload: [String: Any] = ["command_name": "config", "firebase_session_timeout_seconds": "60", "firebase_session_minimum_seconds": "30", "firebase_analytics_enabled": "true", "firebase_log_level": "max"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.createAnalyticsConfigCallCount)
         }
@@ -62,7 +48,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testCreateAnalyticsConfigShouldNotRun() {
         let expect = expectation(description: "firebase config should not run")
         let payload: [String: Any] = ["command_name": "initialize"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.createAnalyticsConfigCallCount)
         }
@@ -85,7 +71,7 @@ class FirebaseTrackerTests: XCTestCase {
                         ]
             , "param_coupon": "summer2020", "param_campaign": "disney"]
         ]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.logEventWithParamsCallCount)
         }
@@ -96,7 +82,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testLogEventWithoutParams() {
         let expect = expectation(description: "log event should not run")
         let payload: [String: Any] = ["command_name": "logevent", "firebase_event_name": "event_level_up"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.logEventWithoutParamsCallCount)
         }
@@ -107,7 +93,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetScreenNameWithScreenValues() {
         let expect = expectation(description: "set screen name should run")
         let payload: [String: Any] = ["command_name": "setscreenname", "firebase_screen_name": "product_view", "firebase_screen_class": "ProductDetailViewController"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.setScreenNameCallCount)
         }
@@ -118,7 +104,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetScreenNameWithoutScreenValues() {
         let expect = expectation(description: "set screen name should not run")
         let payload: [String: Any] = ["command_name": "setscreenname"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.setScreenNameCallCount)
         }
@@ -129,7 +115,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetUserPropertyWithValues() {
         let expect = expectation(description: "set user property should run")
         let payload: [String: Any] = ["command_name": "setuserproperty", "firebase_property_name": "favorite_color", "firebase_property_value": "blue"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.setUserPropertyCallCount)
         }
@@ -140,7 +126,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetUserPropertyWithoutValues() {
         let expect = expectation(description: "set user property should not run")
         let payload: [String: Any] = ["command_name": "setuserproperty"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.setUserPropertyCallCount)
         }
@@ -151,7 +137,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetUserPropertiesWithValues() {
         let expect = expectation(description: "set user property should run")
         let payload: [String: Any] = ["command_name": "setuserproperty", "firebase_property_name": ["favorite_color", "nickname"], "firebase_property_value": ["blue", "sparky"]]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(2, firebaseTracker.setUserPropertyCallCount)
         }
@@ -162,7 +148,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetUserIdWithUserId() {
         let expect = expectation(description: "set user id should run")
         let payload: [String: Any] = ["command_name": "setuserid", "firebase_user_id": "abc123"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.setUserIdCallCount)
         }
@@ -173,7 +159,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetUserIdWithoutUserId() {
         let expect = expectation(description: "set user id should not run")
         let payload: [String: Any] = ["command_name": "setuserid"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.setUserIdCallCount)
         }
@@ -184,7 +170,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testMapParamsWhenTheyDontExistInLookup() {
                 let expect = expectation(description: "set user id should not run")
         let payload: [String: Any] = ["command_name": "logevent", "firebase_event_name": "ecommerce_purchase", "coupon": "couponCode", "currency": "AUD", "value": 19.99, "tax": 1.99, "shipping": 2.00, "transaction_id": "1232312321"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.logEventWithParamsCallCount)
         }
@@ -195,7 +181,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testMapParamsWhenTheyDoExistInLookup() {
                 let expect = expectation(description: "set user id should not run")
         let payload: [String: Any] = ["command_name": "logevent", "firebase_event_name": "event_ecommerce_purchase", "param_coupon": "couponCode", "param_currency": "AUD", "param_value": 19.99, "param_tax": 1.99, "param_shipping": 2.00, "param_transaction_id": "1232312321"]
-        if let response = createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .webview, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.logEventWithParamsCallCount)
         }
@@ -207,7 +193,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testCreateAnalyticsConfigWithoutValuesJSON() {
         let expect = expectation(description: "firebase config should run")
         let payload: [String: Any] = ["command_name": "config"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.createAnalyticsConfigCallCount)
         }
@@ -218,7 +204,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testCreateAnalyticsConfigWithValuesJSON() {
         let expect = expectation(description: "firebase config should run")
         let payload: [String: Any] = ["command_name": "config", "firebase_session_timeout_seconds": "60", "firebase_session_minimum_seconds": "30", "firebase_analytics_enabled": "true", "firebase_log_level": "max"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.createAnalyticsConfigCallCount)
         }
@@ -229,7 +215,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testCreateAnalyticsConfigShouldNotRunJSON() {
         let expect = expectation(description: "firebase config should not run")
         let payload: [String: Any] = ["command_name": "initialize"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.createAnalyticsConfigCallCount)
         }
@@ -252,7 +238,7 @@ class FirebaseTrackerTests: XCTestCase {
                         ]
             , "param_coupon": "summer2020", "param_campaign": "disney"]
         ]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.logEventWithParamsCallCount)
         }
@@ -263,7 +249,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testLogEventWithoutParamsJSON() {
         let expect = expectation(description: "log event should not run")
         let payload: [String: Any] = ["command_name": "logevent", "firebase_event_name": "event_level_up"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.logEventWithoutParamsCallCount)
         }
@@ -274,7 +260,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetScreenNameWithScreenValuesJSON() {
         let expect = expectation(description: "set screen name should run")
         let payload: [String: Any] = ["command_name": "setscreenname", "firebase_screen_name": "product_view", "firebase_screen_class": "ProductDetailViewController"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.setScreenNameCallCount)
         }
@@ -285,7 +271,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetScreenNameWithoutScreenValuesJSON() {
         let expect = expectation(description: "set screen name should not run")
         let payload: [String: Any] = ["command_name": "setscreenname"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.setScreenNameCallCount)
         }
@@ -296,7 +282,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetUserPropertyWithValuesJSON() {
         let expect = expectation(description: "set user property should run")
         let payload: [String: Any] = ["command_name": "setuserproperty", "firebase_property_name": "favorite_color", "firebase_property_value": "blue"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.setUserPropertyCallCount)
         }
@@ -307,7 +293,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetUserPropertyWithoutValuesJSON() {
         let expect = expectation(description: "set user property should not run")
         let payload: [String: Any] = ["command_name": "setuserproperty"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.setUserPropertyCallCount)
         }
@@ -318,7 +304,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetUserPropertiesWithValuesJSON() {
         let expect = expectation(description: "set user property should run")
         let payload: [String: Any] = ["command_name": "setuserproperty", "firebase_property_name": ["favorite_color", "nickname"], "firebase_property_value": ["blue", "sparky"]]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(2, firebaseTracker.setUserPropertyCallCount)
         }
@@ -329,7 +315,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetUserIdWithUserIdJSON() {
         let expect = expectation(description: "set user id should run")
         let payload: [String: Any] = ["command_name": "setuserid", "firebase_user_id": "abc123"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(1, firebaseTracker.setUserIdCallCount)
         }
@@ -340,7 +326,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testSetUserIdWithoutUserIdJSON() {
         let expect = expectation(description: "set user id should not run")
         let payload: [String: Any] = ["command_name": "setuserid"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.setUserIdCallCount)
         }
@@ -351,7 +337,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testMapParamsWhenTheyDontExistInLookupJSON() {
                 let expect = expectation(description: "set user id should not run")
         let payload: [String: Any] = ["command_name": "logevent", "firebase_event_name": "ecommerce_purchase", "coupon": "couponCode", "currency": "AUD", "value": 19.99, "tax": 1.99, "shipping": 2.00, "transaction_id": "1232312321"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.logEventWithParamsCallCount)
         }
@@ -362,7 +348,7 @@ class FirebaseTrackerTests: XCTestCase {
     func testMapParamsWhenTheyDoExistInLookupJSON() {
                 let expect = expectation(description: "set user id should not run")
         let payload: [String: Any] = ["command_name": "logevent", "firebase_event_name": "event_ecommerce_purchase", "param_coupon": "couponCode", "param_currency": "AUD", "param_value": 19.99, "param_tax": 1.99, "param_shipping": 2.00, "param_transaction_id": "1232312321"]
-        if let response = createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
+        if let response = HttpTestHelpers.createRemoteCommandResponse(type: .JSON, commandId: "firebase", payload: payload) {
             firebaseCommand.completion(response)
             XCTAssertEqual(0, firebaseTracker.logEventWithParamsCallCount)
         }
